+++ packages/compat/uni-builder/src/types.ts
@@ -208,6 +208,12 @@ export type UniBuilderExtraConfig = {
      */
     checkSyntax?: boolean | PluginCheckSyntaxOptions;
   };
+  experiments?: {
+    /**
+     * Enable the ability for source code building
+     */
+    sourceBuild?: boolean;
+  };
 };
 
 export type SriOptions = {