+++ packages/shared/src/config.ts
@@ -30,12 +30,11 @@ import type {
   NormalizedToolsConfig,
   NormalizedConfig,
 } from './types';
-import { pick } from './pick';
 import { logger } from './logger';
 import { join } from 'path';
-import { color } from './color';
 import type { minify } from 'terser';
 import fse from '../compiled/fs-extra';
+import { pick, color, upperFirst } from './utils';
 
 import _ from 'lodash';
 import { DEFAULT_DEV_HOST } from './constants';
@@ -172,7 +171,7 @@ export async function outputInspectConfigFiles({
 
       return {
         path: outputFilePath,
-        label: `${_.upperFirst(configType)} Config (${suffix})`,
+        label: `${upperFirst(configType)} Config (${suffix})`,
         content,
       };
     }),