+++ packages/shared/src/formatStats.ts
@@ -85,7 +85,7 @@ function formatMessage(stats: webpack.StatsError | string) {
   return message.trim();
 }
 
-function formatStatsMessages(
+export function formatStatsMessages(
   json?: Pick<StatsCompilation, 'errors' | 'warnings'>,
 ): {
   errors: string[];
@@ -112,5 +112,3 @@ function formatStatsMessages(
 
   return result;
 }
-
-export { formatStatsMessages };