+++ packages/create-rsbuild/CHANGELOG.md
@@ -1,5 +1,7 @@
 # create-rsbuild
 
+## 0.1.5
+
 ## 0.1.4
 
 ### Patch Changes