+++ packages/monorepo-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rsbuild/monorepo-utils
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies [ca87f86f]
+  - @rsbuild/shared@0.1.5
+
 ## 0.1.4
 
 ### Patch Changes